Fixes for feedback received from Brandon:
"We received test orders for Skelly and Memorial.  No store number referenced on either.
Supplier code is gibberish.  What else would you like me to look at?
If possible, it would help us to have the PLU (SKU) and the order quantity next to each other.  Those are
the items we have to key."